+++ src/gui/ldGauge.c
@@ -97,7 +97,7 @@ ldGauge_t* ldGauge_init( ld_scene_t *ptScene,ldGauge_t *ptWidget, uint16_t nameI
     ptWidget->centreOffsetX=centreOffsetX;
     ptWidget->centreOffsetY=centreOffsetY;
     
-    LOG_INFO("[init][gauge] id:%d, size:%llu", nameId,sizeof (*ptWidget));
+    LOG_INFO("[init][gauge] id:%d, size:%d", nameId,(int)sizeof (*ptWidget));
     return ptWidget;
 }
