+++ src/gui/ldText.c
@@ -203,7 +203,7 @@ ldText_t* ldText_init(ld_scene_t *ptScene,ldText_t *ptWidget, uint16_t nameId, u
         ldMsgConnect(ptWidget,SIGNAL_RELEASE,slotTextVerticalScroll);
     }
 
-    LOG_INFO("[init][text] id:%d, size:%llu", nameId,sizeof (*ptWidget));
+    LOG_INFO("[init][text] id:%d, size:%d", nameId,(int)sizeof (*ptWidget));
     return ptWidget;
 }
